requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-wtf==1.2.1
qrcode==7.4.2
fpdf2==2.7.6
Pillow==10.0.0
pycryptodome==3.19.0
werkzeug==2.3.7
email-validator==2.0.0
flask-migrate==4.0.5
python-dotenv==1.0.0
